Creating a fn variable.
This code will not run.

The only bit of code shared between the continueIndicatingFruit*
functions is the indicateFruit* function that's being called. It's
just a value/variable like any other, so it can be made a parameter
of the function.

Doing so will require all functions that this is based off of to send
another argument. That will come later.